requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
argcomplete==3.6.2
arrow==1.3.0
binaryornot==0.4.4
black==25.9.0
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
codecarbon==3.0.5
contourpy==1.3.3
cookiecutter==2.6.0
cookiecutter-data-science==2.3.0
cryptography==46.0.1
cycler==0.12.1
-e git+https://github.com/facebookresearch/detectron2.git@a1ce2f956a1d2212ad672e3c47d53405c2fe4312#egg=detectron2
fief-client==0.20.0
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
fvcore==0.1.5.post20221221
grpcio==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
hydra-core==1.3.2
idna==3.10
iopath==0.1.9
Jinja2==3.1.6
jwcrypto==1.5.6
kiwisolver==1.4.9
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.5
numpy==2.2.6
nvidia-ml-py==13.580.82
omegaconf==2.3.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==11.3.0
pipx==1.7.1
platformdirs==4.4.0
portalocker==3.2.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.32.1
psutil==7.1.0
py-cpuinfo==9.0.0
pycocotools==2.0.10
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pynvml==13.0.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.2
questionary==2.1.1
RapidFuzz==3.14.1
requests==2.32.5
rich==14.1.0
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
text-unidecode==1.3
tomlkit==0.13.3
torch==2.8.0
tqdm==4.67.1
typer==0.19.2
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
userpath==1.9.2
wcwidth==0.2.14
Werkzeug==3.1.3
wheel==0.45.1
yacs==0.1.8
yaspin==3.2.0
